requirements.txt
colorama==0.3.9
requests==2.18.1
xmltodict==0.11.0
tqdm==4.14.0
elizabeth==0.4.0
flask==0.12.2
psycopg2==2.7.1
flake8==3.3.0
flask-autodoc==0.1.2
python-dateutil==2.6.0
pytest-flask==0.10.0
flaky==3.4.0